requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
Markdown==3.7
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
sqlparse==0.5.1
